pyproject.toml
[tool.poetry]
name = "kaggle agent"
version = "0.1.0"
description = ""
authors = ["msnp nematpoor <msnp@duck.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.7"
langgraph = "^0.1.6"
langchain-openai = "^0.1.14"
pandas = "^2.2.2"
ipython = "^8.26.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.7"
langfuse = "^2.39.1"
seaborn = "^0.13.2"
matplotlib = "^3.9.1"
jupyter-client = "^8.6.2"
websocket-client = "^1.8.0"
nbformat = "^5.10.4"
nbconvert = "^7.16.4"
httpx = {extras = ["socks"], version = "^0.27.0"}
grandalf = "^0.8"
black = "^24.4.2"
orjson = "^3.10.6"
loguru = "^0.7.2"
pyyaml = "^6.0.1"
streamlit = "^1.37.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
